requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.4
google_search_results==2.4.2
newsapi_python==0.2.7
openai==1.54.3
python-dotenv==1.0.1
Requests==2.32.3
serpapi==0.1.5
typing_extensions==4.12.2
websockets==14.0
